rename: ast_access -> workspace (since there will be loggin)
