+++ dbuf-lsp/src/core/mod.rs
@@ -1,11 +1,11 @@
 //! Helpfull tools.
 
-pub mod ast_access;
 pub mod ast_visitor;
 pub mod dbuf_language;
 pub mod errors;
 pub mod navigator;
 pub mod pretty_printer;
+pub mod workspace;
 
 // might be deleted in future
 mod default_ast;
